feat: implement global webinar banner using theme override
- Replaced individual LayoutWithBanner usage with theme Layout override
- Banner now automatically appears on all pages except webinar pages
- Fixed padding to 54px to match actual banner height
- Cleaner architecture with single source of truth for banner
